resolve duplicate code
