+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/Claim.java
@@ -277,7 +277,7 @@ public class Claim {
     public Optional<String> getPreferredDQCourt() {
         return Optional.ofNullable(preferredDQCourt);
     }
-    
+
     @Override
     public String toString() {
         return ReflectionToStringBuilder.toString(this, ourStyle());
