+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/legaladvisor/DrawJudgeOrderCallbackHandler.java
@@ -2,66 +2,28 @@ package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.legaladvisor;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
-import uk.gov.hmcts.cmc.ccd.domain.CCDClaimDocument;
-import uk.gov.hmcts.cmc.ccd.domain.CCDCollectionElement;
-import uk.gov.hmcts.cmc.ccd.domain.CCDDirectionOrder;
-import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
-import uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
-import uk.gov.hmcts.cmc.ccd.domain.claimantresponse.CCDResponseRejection;
-import uk.gov.hmcts.cmc.ccd.domain.defendant.CCDRespondent;
-import uk.gov.hmcts.cmc.ccd.domain.directionsquestionnaire.CCDDirectionsQuestionnaire;
-import uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDResponseSubjectType;
-import uk.gov.hmcts.cmc.claimstore.exceptions.CallbackException;
-import uk.gov.hmcts.cmc.claimstore.services.LegalOrderGenerationDeadlinesCalculator;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.DocAssemblyService;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.rules.GenerateOrderRule;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.legaladvisor.HearingCourt;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.legaladvisor.HearingCourtDetailsFinder;
 import uk.gov.hmcts.cmc.claimstore.services.notifications.legaladvisor.OrderDrawnNotificationService;
 import uk.gov.hmcts.cmc.claimstore.services.staff.content.legaladvisor.LegalOrderService;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.cmc.claimstore.utils.DirectionsQuestionnaireUtils;
-import uk.gov.hmcts.cmc.domain.models.Claim;
-import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
-import uk.gov.hmcts.reform.docassembly.domain.DocAssemblyResponse;
 
 import java.time.Clock;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
-import static uk.gov.hmcts.cmc.ccd.domain.CCDClaimDocumentType.ORDER_DIRECTIONS;
-import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.NO;
-import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.YES;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.DRAW_JUDGES_ORDER;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDDirectionPartyType.BOTH;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDOrderDirectionType.DOCUMENTS;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDOrderDirectionType.EYEWITNESS;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.JUDGE;
-import static uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams.Params.BEARER_TOKEN;
-import static uk.gov.hmcts.cmc.domain.utils.LocalDateTimeFactory.UTC_ZONE;
-import static uk.gov.hmcts.cmc.domain.utils.LocalDateTimeFactory.nowInUTC;
 
 @Service
 @ConditionalOnProperty(prefix = "doc_assembly", name = "url")
@@ -70,28 +32,14 @@ public class DrawJudgeOrderCallbackHandler extends CallbackHandler {
 
     private static final List<Role> ROLES = ImmutableList.of(JUDGE);
     private static final List<CaseEvent> EVENTS = Collections.singletonList(DRAW_JUDGES_ORDER);
-    private static final String DOC_UPLOAD_DEADLINE = "docUploadDeadline";
-    private static final String EYEWITNESS_UPLOAD_DEADLINE = "eyewitnessUploadDeadline";
-    private static final String DOC_UPLOAD_FOR_PARTY = "docUploadForParty";
-    private static final String EYEWITNESS_UPLOAD_FOR_PARTY = "eyewitnessUploadForParty";
-    private static final String PAPER_DETERMINATION = "paperDetermination";
-    private static final String DRAFT_ORDER_DOC = "draftOrderDoc";
-    private static final String NEW_REQUESTED_COURT = "newRequestedCourt";
-    private static final String PREFERRED_COURT_OBJECTING_PARTY = "preferredCourtObjectingParty";
-    private static final String PREFERRED_COURT_OBJECTING_REASON = "preferredCourtObjectingReason";
-    private static final String DIRECTION_LIST = "directionList";
-    private static final String PREFERRED_DQ_COURT = "preferredDQCourt";
-    private static final String EXPERT_PERMISSION_BY_CLAIMANT = "expertReportPermissionPartyAskedByClaimant";
-    private static final String EXPERT_PERMISSION_BY_DEFENDANT = "expertReportPermissionPartyAskedByDefendant";
 
     private final Clock clock;
     private final OrderDrawnNotificationService orderDrawnNotificationService;
     private final CaseDetailsConverter caseDetailsConverter;
     private final LegalOrderService legalOrderService;
     private final HearingCourtDetailsFinder hearingCourtDetailsFinder;
-    private final DocAssemblyService docAssemblyService;
-    private final LegalOrderGenerationDeadlinesCalculator legalOrderGenerationDeadlinesCalculator;
-    private final GenerateOrderRule generateOrderRule;
+    private final OrderCreator orderCreator;
+    private final OrderPostProcessor orderPostProcessor;
 
     @Autowired
     public DrawJudgeOrderCallbackHandler(
@@ -100,18 +48,16 @@ public class DrawJudgeOrderCallbackHandler extends CallbackHandler {
         CaseDetailsConverter caseDetailsConverter,
         LegalOrderService legalOrderService,
         HearingCourtDetailsFinder hearingCourtDetailsFinder,
-        DocAssemblyService docAssemblyService,
-        LegalOrderGenerationDeadlinesCalculator legalOrderGenerationDeadlinesCalculator,
-        GenerateOrderRule generateOrderRule
+        OrderCreator orderCreator,
+        OrderPostProcessor orderPostProcessor
     ) {
         this.clock = clock;
         this.orderDrawnNotificationService = orderDrawnNotificationService;
         this.caseDetailsConverter = caseDetailsConverter;
         this.legalOrderService = legalOrderService;
         this.hearingCourtDetailsFinder = hearingCourtDetailsFinder;
-        this.docAssemblyService = docAssemblyService;
-        this.legalOrderGenerationDeadlinesCalculator = legalOrderGenerationDeadlinesCalculator;
-        this.generateOrderRule = generateOrderRule;
+        this.orderCreator = orderCreator;
+        this.orderPostProcessor = orderPostProcessor;
     }
 
     @Override
@@ -127,179 +73,10 @@ public class DrawJudgeOrderCallbackHandler extends CallbackHandler {
     @Override
     protected Map<CallbackType, Callback> callbacks() {
         return ImmutableMap.of(
-            CallbackType.ABOUT_TO_START, this::prepopulateOrder,
-            CallbackType.MID, this::generateOrder,
-            CallbackType.ABOUT_TO_SUBMIT, this::copyDraftToCaseDocument,
-            CallbackType.SUBMITTED, this::notifyPartiesAndPrintOrder
+            CallbackType.ABOUT_TO_START, orderCreator::prepopulateOrder,
+            CallbackType.MID, orderCreator::generateOrder,
+            CallbackType.ABOUT_TO_SUBMIT, orderPostProcessor::copyDraftToCaseDocument,
+            CallbackType.SUBMITTED, orderPostProcessor::notifyPartiesAndPrintOrder
         );
-    }
-
-    private CallbackResponse prepopulateOrder(CallbackParams callbackParams) {
-        logger.info("Generate order callback: pre populating order fields");
-        CallbackRequest callbackRequest = callbackParams.getRequest();
-
-        Map<String, Object> data = new HashMap<>();
-        data.put(DIRECTION_LIST, ImmutableList.of(DOCUMENTS.name(), EYEWITNESS.name()));
-        CaseDetails caseDetails = callbackRequest.getCaseDetails();
-        Claim claim = caseDetailsConverter.extractClaim(caseDetails);
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(caseDetails);
-        addCourtData(claim, ccdCase, data);
-
-        LocalDate deadline = legalOrderGenerationDeadlinesCalculator.calculateOrderGenerationDeadlines();
-        data.put(DOC_UPLOAD_DEADLINE, deadline);
-        data.put(EYEWITNESS_UPLOAD_DEADLINE, deadline);
-        data.put(DOC_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(EYEWITNESS_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(PAPER_DETERMINATION, NO.name());
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(data)
-            .build();
-    }
-
-    private CallbackResponse generateOrder(CallbackParams callbackParams) {
-        CaseDetails caseDetails = callbackParams.getRequest().getCaseDetails();
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(caseDetails);
-
-        List<String> validations = generateOrderRule.validateExpectedFieldsAreSelectedByLegalAdvisor(ccdCase);
-        if (!validations.isEmpty()) {
-            return AboutToStartOrSubmitCallbackResponse.builder().errors(validations).build();
-        }
-
-        String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
-        DocAssemblyResponse docAssemblyResponse = docAssemblyService.createOrder(ccdCase, authorisation);
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(ImmutableMap.of(
-                DRAFT_ORDER_DOC,
-                CCDDocument.builder().documentUrl(docAssemblyResponse.getRenditionOutputLocation()).build()
-            ))
-            .build();
-    }
-
-    private CallbackResponse notifyPartiesAndPrintOrder(CallbackParams callbackParams) {
-        CaseDetails caseDetails = callbackParams.getRequest().getCaseDetails();
-        Claim claim = caseDetailsConverter.extractClaim(caseDetails);
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(caseDetails);
-        notifyParties(claim);
-        String authorisation = callbackParams.getParams().get(BEARER_TOKEN).toString();
-        return printOrder(authorisation, claim, ccdCase);
-    }
-
-    private void notifyParties(Claim claim) {
-        orderDrawnNotificationService.notifyClaimant(claim);
-        orderDrawnNotificationService.notifyDefendant(claim);
-    }
-
-    private CallbackResponse printOrder(String authorisation, Claim claim, CCDCase ccdCase) {
-        CCDDocument draftOrderDoc = ccdCase.getDraftOrderDoc();
-        legalOrderService.print(authorisation, claim, draftOrderDoc);
-        return SubmittedCallbackResponse.builder().build();
-    }
-
-    private CallbackResponse copyDraftToCaseDocument(CallbackParams callbackParams) {
-        CallbackRequest callbackRequest = callbackParams.getRequest();
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
-
-        CCDDocument draftOrderDoc = Optional.ofNullable(ccdCase.getDraftOrderDoc())
-            .orElseThrow(() -> new CallbackException("Draft order not present"));
-
-        HearingCourt hearingCourt = Optional.ofNullable(ccdCase.getHearingCourt())
-            .map(hearingCourtDetailsFinder::findHearingCourtAddress)
-            .orElseGet(() -> HearingCourt.builder().build());
-
-        CCDCase updatedCase = ccdCase.toBuilder()
-            .caseDocuments(updateCaseDocumentsWithOrder(ccdCase, draftOrderDoc))
-            .directionOrder(CCDDirectionOrder.builder()
-                .createdOn(nowInUTC())
-                .hearingCourtName(hearingCourt.getName())
-                .hearingCourtAddress(hearingCourt.getAddress())
-                .build())
-            .build();
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(caseDetailsConverter.convertToMap(updatedCase))
-            .build();
-    }
-
-    private List<CCDCollectionElement<CCDClaimDocument>> updateCaseDocumentsWithOrder(
-        CCDCase ccdCase,
-        CCDDocument draftOrderDoc
-    ) {
-        CCDCollectionElement<CCDClaimDocument> claimDocument = CCDCollectionElement.<CCDClaimDocument>builder()
-            .value(CCDClaimDocument.builder()
-                .documentLink(draftOrderDoc)
-                .documentName(draftOrderDoc.getDocumentFileName())
-                .createdDatetime(LocalDateTime.now(clock.withZone(UTC_ZONE)))
-                .documentType(ORDER_DIRECTIONS)
-                .build())
-            .build();
-
-        return ImmutableList.<CCDCollectionElement<CCDClaimDocument>>builder()
-            .addAll(ccdCase.getCaseDocuments())
-            .add(claimDocument)
-            .build();
-    }
-
-    private void addCourtData(Claim claim, CCDCase ccdCase, Map<String, Object> data) {
-        CCDRespondent respondent = ccdCase.getRespondents().get(0).getValue();
-
-        CCDResponseRejection claimantResponse = Optional.ofNullable(respondent.getClaimantResponse())
-            .map(r -> (CCDResponseRejection) r)
-            .orElseThrow(() -> new IllegalStateException("Claimant Response not present"));
-
-        CCDDirectionsQuestionnaire claimantDQ = claimantResponse.getDirectionsQuestionnaire();
-        CCDDirectionsQuestionnaire defendantDQ = respondent.getDirectionsQuestionnaire();
-
-        String newRequestedCourt = null;
-        String preferredCourtObjectingParty = null;
-        String preferredCourtObjectingReason = null;
-
-        if (claimantDQ != null && StringUtils.isNotBlank(claimantDQ.getExceptionalCircumstancesReason())) {
-            newRequestedCourt = claimantDQ.getHearingLocation();
-            preferredCourtObjectingParty = CCDResponseSubjectType.RES_CLAIMANT.getValue();
-            preferredCourtObjectingReason = claimantDQ.getExceptionalCircumstancesReason();
-        } else if (defendantDQ != null && StringUtils.isNotBlank(
-            defendantDQ.getExceptionalCircumstancesReason())) {
-            newRequestedCourt = defendantDQ.getHearingLocation();
-            preferredCourtObjectingParty = CCDResponseSubjectType.RES_DEFENDANT.getValue();
-            preferredCourtObjectingReason = defendantDQ.getExceptionalCircumstancesReason();
-        }
-
-        data.put(NEW_REQUESTED_COURT, newRequestedCourt);
-        data.put(PREFERRED_COURT_OBJECTING_PARTY, preferredCourtObjectingParty);
-        data.put(PREFERRED_COURT_OBJECTING_REASON, preferredCourtObjectingReason);
-        data.put(PREFERRED_DQ_COURT, DirectionsQuestionnaireUtils.getPreferredCourt(claim));
-
-        if (Optional.ofNullable(claimantDQ).isPresent()) {
-            data.put(EXPERT_PERMISSION_BY_CLAIMANT, hasRequestedExpertPermission(claimantDQ));
-        }
-
-        if (Optional.ofNullable(defendantDQ).isPresent()) {
-            data.put(EXPERT_PERMISSION_BY_DEFENDANT, hasRequestedExpertPermission(defendantDQ));
-        }
-    }
-
-    private CCDYesNoOption hasRequestedExpertPermission(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getExpertRequired() != null
-            && directionsQuestionnaire.getExpertRequired().toBoolean()
-            && (hasRequestedForPermissionWithProvidedEvidence(directionsQuestionnaire)
-            || hasProvidedExpertReports(directionsQuestionnaire))
-            ? YES
-            : NO;
-    }
-
-    private boolean hasProvidedExpertReports(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getExpertReports() != null
-            && !directionsQuestionnaire.getExpertReports().isEmpty();
-    }
-
-    private boolean hasRequestedForPermissionWithProvidedEvidence(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getPermissionForExpert() != null
-            && directionsQuestionnaire.getPermissionForExpert().toBoolean()
-            && StringUtils.isNotBlank(directionsQuestionnaire.getExpertEvidenceToExamine());
     }
 }
