+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/legaladvisor/GenerateOrderCallbackHandler.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.legaladvisor;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,44 +9,23 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
-import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
-import uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
-import uk.gov.hmcts.cmc.ccd.domain.claimantresponse.CCDResponseRejection;
-import uk.gov.hmcts.cmc.ccd.domain.defendant.CCDRespondent;
-import uk.gov.hmcts.cmc.ccd.domain.directionsquestionnaire.CCDDirectionsQuestionnaire;
-import uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDResponseSubjectType;
 import uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights;
-import uk.gov.hmcts.cmc.claimstore.services.LegalOrderGenerationDeadlinesCalculator;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.DocAssemblyService;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.Role;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.Callback;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackHandler;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackType;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.rules.GenerateOrderRule;
 import uk.gov.hmcts.cmc.claimstore.utils.CaseDetailsConverter;
-import uk.gov.hmcts.cmc.claimstore.utils.DirectionsQuestionnaireUtils;
-import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
-import uk.gov.hmcts.reform.docassembly.domain.DocAssemblyResponse;
 
-import java.time.LocalDate;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
-import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.NO;
-import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.YES;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.ACTION_REVIEW_COMMENTS;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.GENERATE_ORDER;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDDirectionPartyType.BOTH;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDOrderDirectionType.DOCUMENTS;
-import static uk.gov.hmcts.cmc.ccd.domain.legaladvisor.CCDOrderDirectionType.EYEWITNESS;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.DRAFTED_BY_LEGAL_ADVISOR;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.LEGAL_ADVISOR;
@@ -73,36 +51,31 @@ public class GenerateOrderCallbackHandler extends CallbackHandler {
     private static final String EXPERT_PERMISSION_BY_DEFENDANT = "expertReportPermissionPartyAskedByDefendant";
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
+    private final OrderCreator orderCreator;
 
     @Value("${doc_assembly.templateId}")
     private String templateId;
 
-    private final LegalOrderGenerationDeadlinesCalculator legalOrderGenerationDeadlinesCalculator;
     private final CaseDetailsConverter caseDetailsConverter;
-    private final DocAssemblyService docAssemblyService;
     private final AppInsights appInsights;
-    private final GenerateOrderRule generateOrderRule;
 
     @Autowired
     public GenerateOrderCallbackHandler(
-        LegalOrderGenerationDeadlinesCalculator legalOrderGenerationDeadlinesCalculator,
+        OrderCreator orderCreator,
         CaseDetailsConverter caseDetailsConverter,
-        DocAssemblyService docAssemblyService,
-        AppInsights appInsights,
-        GenerateOrderRule generateOrderRule
+        AppInsights appInsights
     ) {
-        this.legalOrderGenerationDeadlinesCalculator = legalOrderGenerationDeadlinesCalculator;
+        this.orderCreator = orderCreator;
         this.caseDetailsConverter = caseDetailsConverter;
-        this.docAssemblyService = docAssemblyService;
         this.appInsights = appInsights;
-        this.generateOrderRule = generateOrderRule;
     }
 
     @Override
     protected Map<CallbackType, Callback> callbacks() {
         return ImmutableMap.of(
-            CallbackType.ABOUT_TO_START, this::prepopulateOrder,
-            CallbackType.MID, this::generateOrder
+            CallbackType.ABOUT_TO_START, orderCreator::prepopulateOrder,
+            CallbackType.MID, orderCreator::generateOrder,
+            CallbackType.SUBMITTED, this::raiseAppInsight
         );
     }
 
@@ -116,111 +89,10 @@ public class GenerateOrderCallbackHandler extends CallbackHandler {
         return ROLES;
     }
 
-    private CallbackResponse prepopulateOrder(CallbackParams callbackParams) {
-        logger.info("Generate order callback: pre populating order fields");
-        CallbackRequest callbackRequest = callbackParams.getRequest();
-
-        Map<String, Object> data = new HashMap<>();
-        data.put(DIRECTION_LIST, ImmutableList.of(DOCUMENTS.name(), EYEWITNESS.name()));
-
-        Claim claim = caseDetailsConverter.extractClaim(callbackRequest.getCaseDetails());
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
-        addCourtData(claim, ccdCase, data);
-
-        LocalDate deadline = legalOrderGenerationDeadlinesCalculator.calculateOrderGenerationDeadlines();
-        data.put(DOC_UPLOAD_DEADLINE, deadline);
-        data.put(EYEWITNESS_UPLOAD_DEADLINE, deadline);
-        data.put(DOC_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(EYEWITNESS_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(PAPER_DETERMINATION, NO.name());
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(data)
-            .build();
-    }
-
-    private CallbackResponse generateOrder(CallbackParams callbackParams) {
-        logger.info("Generate order callback: creating order document");
-        CallbackRequest callbackRequest = callbackParams.getRequest();
-        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
-
-        List<String> validations = generateOrderRule.validateExpectedFieldsAreSelectedByLegalAdvisor(ccdCase);
-        if (!validations.isEmpty()) {
-            return AboutToStartOrSubmitCallbackResponse.builder().errors(validations).build();
-        }
-
-        String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
-        DocAssemblyResponse docAssemblyResponse = docAssemblyService.createOrder(ccdCase, authorisation);
-        logger.info("Generate order callback: received response from doc assembly");
-
+    private CallbackResponse raiseAppInsight(CallbackParams callbackParams) {
+        logger.info("Generate order callback: raise app insight");
+        CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackParams.getRequest().getCaseDetails());
         appInsights.trackEvent(DRAFTED_BY_LEGAL_ADVISOR, REFERENCE_NUMBER, ccdCase.getPreviousServiceCaseReference());
-
-        return AboutToStartOrSubmitCallbackResponse
-            .builder()
-            .data(ImmutableMap.of(
-                DRAFT_ORDER_DOC,
-                CCDDocument.builder().documentUrl(docAssemblyResponse.getRenditionOutputLocation()).build()
-            ))
-            .build();
-    }
-
-    private void addCourtData(Claim claim, CCDCase ccdCase, Map<String, Object> data) {
-        CCDRespondent respondent = ccdCase.getRespondents().get(0).getValue();
-
-        CCDResponseRejection claimantResponse = Optional.ofNullable(respondent.getClaimantResponse())
-            .map(r -> (CCDResponseRejection) r)
-            .orElseThrow(() -> new IllegalStateException("Claimant Response not present"));
-
-        CCDDirectionsQuestionnaire claimantDQ = claimantResponse.getDirectionsQuestionnaire();
-        CCDDirectionsQuestionnaire defendantDQ = respondent.getDirectionsQuestionnaire();
-
-        String newRequestedCourt = null;
-        String preferredCourtObjectingParty = null;
-        String preferredCourtObjectingReason = null;
-
-        if (claimantDQ != null && StringUtils.isNotBlank(claimantDQ.getExceptionalCircumstancesReason())) {
-            newRequestedCourt = claimantDQ.getHearingLocation();
-            preferredCourtObjectingParty = CCDResponseSubjectType.RES_CLAIMANT.getValue();
-            preferredCourtObjectingReason = claimantDQ.getExceptionalCircumstancesReason();
-        } else if (defendantDQ != null && StringUtils.isNotBlank(
-            defendantDQ.getExceptionalCircumstancesReason())) {
-            newRequestedCourt = defendantDQ.getHearingLocation();
-            preferredCourtObjectingParty = CCDResponseSubjectType.RES_DEFENDANT.getValue();
-            preferredCourtObjectingReason = defendantDQ.getExceptionalCircumstancesReason();
-        }
-
-        data.put(NEW_REQUESTED_COURT, newRequestedCourt);
-        data.put(PREFERRED_COURT_OBJECTING_PARTY, preferredCourtObjectingParty);
-        data.put(PREFERRED_COURT_OBJECTING_REASON, preferredCourtObjectingReason);
-        data.put(PREFERRED_DQ_COURT, DirectionsQuestionnaireUtils.getPreferredCourt(claim));
-
-        if (Optional.ofNullable(claimantDQ).isPresent()) {
-            data.put(EXPERT_PERMISSION_BY_CLAIMANT, hasRequestedExpertPermission(claimantDQ));
-        }
-
-        if (Optional.ofNullable(defendantDQ).isPresent()) {
-            data.put(EXPERT_PERMISSION_BY_DEFENDANT, hasRequestedExpertPermission(defendantDQ));
-        }
-    }
-
-    private CCDYesNoOption hasRequestedExpertPermission(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getExpertRequired() != null
-            && directionsQuestionnaire.getExpertRequired().toBoolean()
-            && (hasRequestedForPermissionWithProvidedEvidence(directionsQuestionnaire)
-            || hasProvidedExpertReports(directionsQuestionnaire))
-            ? YES
-            : NO;
-    }
-
-    private boolean hasProvidedExpertReports(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getExpertReports() != null
-            && !directionsQuestionnaire.getExpertReports().isEmpty();
-    }
-
-    private boolean hasRequestedForPermissionWithProvidedEvidence(CCDDirectionsQuestionnaire directionsQuestionnaire) {
-        return directionsQuestionnaire.getPermissionForExpert() != null
-            && directionsQuestionnaire.getPermissionForExpert().toBoolean()
-            && StringUtils.isNotBlank(directionsQuestionnaire.getExpertEvidenceToExamine());
+        return AboutToStartOrSubmitCallbackResponse.builder().build();
     }
 }
