+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/legaladvisor/GenerateOrderCallbackHandlerTest.java
@@ -43,10 +43,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.data.MapEntry.entry;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.NO;
 import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.YES;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.GENERATE_ORDER;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.DRAFTED_BY_LEGAL_ADVISOR;
 
 @RunWith(MockitoJUnitRunner.class)
 public class GenerateOrderCallbackHandlerTest {
@@ -75,12 +78,11 @@ public class GenerateOrderCallbackHandlerTest {
 
     @Before
     public void setUp() {
-        generateOrderCallbackHandler = new GenerateOrderCallbackHandler(
-            legalOrderGenerationDeadlinesCalculator,
-            caseDetailsConverter,
-            docAssemblyService,
-            appInsights,
-            new GenerateOrderRule());
+        OrderCreator orderCreator = new OrderCreator(legalOrderGenerationDeadlinesCalculator, caseDetailsConverter,
+            docAssemblyService, new GenerateOrderRule());
+
+        generateOrderCallbackHandler = new GenerateOrderCallbackHandler(orderCreator,
+            caseDetailsConverter, appInsights);
 
         ReflectionTestUtils.setField(generateOrderCallbackHandler, "templateId", "testTemplateId");
         ccdCase = SampleData.getCCDCitizenCase(Collections.emptyList());
@@ -98,9 +100,7 @@ public class GenerateOrderCallbackHandlerTest {
                 ).build();
         when(caseDetailsConverter.extractClaim(any(CaseDetails.class))).thenReturn(claim);
         when(caseDetailsConverter.extractCCDCase(any(CaseDetails.class))).thenReturn(ccdCase);
-
-        when(legalOrderGenerationDeadlinesCalculator.calculateOrderGenerationDeadlines())
-            .thenReturn(DEADLINE);
+        when(legalOrderGenerationDeadlinesCalculator.calculateOrderGenerationDeadlines()).thenReturn(DEADLINE);
 
         callbackRequest = CallbackRequest
             .builder()
@@ -377,12 +377,25 @@ public class GenerateOrderCallbackHandlerTest {
     @Test(expected = CallbackException.class)
     public void shouldThrowIfUnimplementedCallbackForValidEvent() {
         CallbackParams callbackParams = CallbackParams.builder()
+            .type(CallbackType.ABOUT_TO_SUBMIT)
+            .request(callbackRequest)
+            .params(ImmutableMap.of(CallbackParams.Params.BEARER_TOKEN, BEARER_TOKEN))
+            .build();
+
+        generateOrderCallbackHandler.handle(callbackParams);
+    }
+
+    @Test
+    public void shouldRaiseAppInsight() {
+        CallbackParams callbackParams = CallbackParams.builder()
             .type(CallbackType.SUBMITTED)
             .request(callbackRequest)
             .params(ImmutableMap.of(CallbackParams.Params.BEARER_TOKEN, BEARER_TOKEN))
             .build();
 
-        generateOrderCallbackHandler
-            .handle(callbackParams);
+        generateOrderCallbackHandler.handle(callbackParams);
+
+        verify(appInsights)
+            .trackEvent(DRAFTED_BY_LEGAL_ADVISOR, REFERENCE_NUMBER, ccdCase.getPreviousServiceCaseReference());
     }
 }
